doc: skipped records must be valid
Plain --skip can just skip input lines without parsing them.

In -b mode, this does not work:
to determine where the record ends (might be on some other line!),
the input module has to actually read the input.
Formatting errors in skipped lines will thus terminate the program
(or, if used with the -e option, leave the parser in an unexpected state).