restrict --trim=fields to --fixed-with mode
It is unclear how field trimming
should relate to quoted fields with whitespace.
Consider this example:

     "  fieldval  "     ;secondfield

Should only the whitespace outside of the quoting characters be removed,
or should the whitespace inside the quotes be removed as well?

Fixed-width files don't use any field quoting,
so the question is moot for them.